feat: optimize stores setup
