+++ front-end/src/renderer/composables/user/useAfterOrganizationSelection.ts
@@ -1,10 +1,7 @@
 import useUserStore from '@renderer/stores/storeUser';
-import useContactsStore from '@renderer/stores/storeContacts';
-import useNotificationsStore from '@renderer/stores/storeNotifications';
-import useWebsocketConnection from '@renderer/stores/storeWebsocketConnection';
 
 import { useRouter } from 'vue-router';
-import { useToast } from 'vue-toast-notification';
+import useSetupStores from '@renderer/composables/user/useSetupStores';
 
 import { get as getStoredMnemonics } from '@renderer/services/mnemonicService';
 
@@ -21,13 +18,10 @@ import {
 export default function useAfterOrganizationSelection() {
   /* Stores */
   const user = useUserStore();
-  const contacts = useContactsStore();
-  const notifications = useNotificationsStore();
-  const ws = useWebsocketConnection();
 
   /* Composables */
   const router = useRouter();
-  const toast = useToast();
+  const setupStores = useSetupStores();
 
   /* Functions */
   const handleStates = async () => {
@@ -46,11 +40,6 @@ export default function useAfterOrganizationSelection() {
     user.mnemonics = mnemonics;
 
     return { keyPairs, mnemonics };
-  };
-
-  const setupStores = async () => {
-    const results = await Promise.allSettled([contacts.fetch(), notifications.setup(), ws.setup()]);
-    results.forEach(r => r.status === 'rejected' && toast.error(r.reason));
   };
 
   const handleNavigation = async () => {
